+++ src/command_modules/azure-cli-resource/azure/cli/command_modules/resource/_help.py
@@ -160,6 +160,18 @@ helps['resource tag'] = """
             az resource tag --tags vmlist=vm1 --id /subscriptions/0b1f6471-1bf0-4dda-aec3-111111111111/resourceGroups/MyResourceGroup/providers/Microsoft.Web/sites/MyWebapp
 """
 
+helps['resource create'] = """
+    type: command
+    short-summary: create a resource.
+    examples:
+       - name: Create a resource by providing a full resource object json. Note, you can also use "@<file>" to load from a json file.
+         text: >
+            az resource create -g myRG -n myPlan --resource-type Microsoft.web/serverFarms --is-full-object --properties "{ \\"location\\":\\"westus\\",\\"sku\\":{\\"name\\":\\"B1\\",\\"tier\\":\\"BASIC\\"}}"
+       - name: Create a resource by only providing resource properties
+         text: >
+            az resource create -g myRG -n myWeb --resource-type Microsoft.web/sites --properties "{\\"serverFarmId\\":\\"myPlan\\"}"
+"""
+
 helps['resource update'] = """
     type: command
     short-summary: Update a resource.
@@ -261,6 +273,18 @@ helps['provider register'] = """
 helps['provider unregister'] = """
     type: command
     short-summary: Unregister a provider.
+"""
+helps['provider operation'] = """
+    type: group
+    short-summary: Get provider operations metadatas.
+"""
+helps['provider operation show'] = """
+    type: command
+    short-summary: Get an individual provider's operations.
+"""
+helps['provider operation list'] = """
+    type: command
+    short-summary: Get operations from all providers.
 """
 helps['tag'] = """
     type: group