+++ src/command_modules/azure-cli-resource/azure/cli/command_modules/resource/custom.py
@@ -27,7 +27,8 @@ from azure.cli.core.profiles import get_sdk, ResourceType
 from ._client_factory import (_resource_client_factory,
                               _resource_policy_client_factory,
                               _resource_lock_client_factory,
-                              _resource_links_client_factory)
+                              _resource_links_client_factory,
+                              _authorization_management_client)
 
 logger = azlogging.get_az_logger(__name__)
 
@@ -208,6 +209,14 @@ def export_deployment_as_template(resource_group_name, deployment_name):
     result = smc.deployments.export_template(resource_group_name, deployment_name)
     print(json.dumps(result.template, indent=2))#pylint: disable=no-member
 
+def create_resource(properties, resource_group_name=None, resource_provider_namespace=None,
+                    parent_resource_path=None, resource_type=None, resource_name=None,
+                    resource_id=None, api_version=None, location=None, is_full_object=False):
+    res = _ResourceUtils(resource_group_name, resource_provider_namespace,
+                         parent_resource_path, resource_type, resource_name,
+                         resource_id, api_version)
+    return res.create_resource(properties, location, is_full_object)
+
 def show_resource(resource_group_name=None, resource_provider_namespace=None,
                   parent_resource_path=None, resource_type=None, resource_name=None,
                   resource_id=None, api_version=None):
@@ -343,6 +352,26 @@ def _update_provider(namespace, registering):
     action = 'Registering' if registering else 'Unregistering'
     msg_template = '%s is still on-going. You can monitor using \'az provider show -n %s\''
     logger.warning(msg_template, action, namespace)
+
+
+def list_provider_operations(api_version=None):
+    api_version = api_version or _get_auth_provider_latest_api_version()
+    auth_client = _authorization_management_client()
+    return auth_client.provider_operations_metadata.list(api_version)
+
+
+def show_provider_operations(resource_provider_namespace, api_version=None):
+    api_version = api_version or _get_auth_provider_latest_api_version()
+    auth_client = _authorization_management_client()
+    return auth_client.provider_operations_metadata.get(resource_provider_namespace, api_version)
+
+
+def _get_auth_provider_latest_api_version():
+    rcf = _resource_client_factory()
+    api_version = _ResourceUtils.resolve_api_version(rcf, 'Microsoft.Authorization',
+                                                     None, 'providerOperations')
+    return api_version
+
 
 def move_resource(ids, destination_group, destination_subscription_id=None):
     '''Moves resources from one resource group to another(can be under different subscription)
@@ -528,6 +557,63 @@ def list_locks(resource_group_name=None, resource_provider_namespace=None,
         resource_group_name, resource_provider_namespace, parent_resource_path, resource_type,
         resource_name, filter=filter_string)
 
+def _validate_lock_params_match_lock(
+        lock_client, name, resource_group_name, resource_provider_namespace,
+        parent_resource_path, resource_type, resource_name):
+    '''
+    Locks are scoped to subscription, resource group or resource.
+    However, the az list command returns all locks for the current scopes
+    and all lower scopes (e.g. resource group level also includes resource locks).
+    This can lead to a confusing user experience where the user specifies a lock
+    name and assumes that it will work, even if they haven't given the right
+    scope. This function attempts to validate the parameters and help the
+    user find the right scope, by first finding the lock, and then infering
+    what it's parameters should be.
+    '''
+    locks = lock_client.management_locks.list_at_subscription_level()
+    found_count = 0 # locks at different levels can have the same name
+    lock_resource_id = None
+    for lock in locks:
+        if lock.name == name:
+            found_count = found_count + 1
+            lock_resource_id = lock.id
+    if found_count == 1:
+        # If we only found one lock, let's validate that the parameters are correct,
+        # if we found more than one, we'll assume the user knows what they're doing
+        # TODO: Add validation for that case too?
+        resource = parse_resource_id(lock_resource_id)
+        _resource_group = resource.get('resource_group', None)
+        _resource_namespace = resource.get('namespace', None)
+        if _resource_group is None:
+            return
+        if resource_group_name != _resource_group:
+            raise CLIError(
+                'Unexpected --resource-group for lock {}, expected {}'.format(
+                    name, _resource_group))
+        if _resource_namespace is None:
+            return
+        if resource_provider_namespace != _resource_namespace:
+            raise CLIError(
+                'Unexpected --namespace for lock {}, expected {}'.format(name, _resource_namespace))
+        if resource.get('grandchild_type', None) is None:
+            _resource_type = resource.get('type', None)
+            _resource_name = resource.get('name', None)
+        else:
+            _resource_type = resource.get('child_type', None)
+            _resource_name = resource.get('child_name', None)
+            parent = (resource['type'] + '/' +  resource['name'])
+            if parent != parent_resource_path:
+                raise CLIError(
+                    'Unexpected --parent for lock {}, expected {}'.format(
+                        name, parent))
+        if resource_type != _resource_type:
+            raise CLIError('Unexpected --resource-type for lock {}, expected {}'.format(
+                name, _resource_type))
+        if resource_name != _resource_name:
+            raise CLIError('Unexpected --resource-name for lock {}, expected {}'.format(
+                name, _resource_name))
+
+
 def get_lock(name, resource_group_name=None):
     '''
     :param name: Name of the lock.
@@ -555,10 +641,15 @@ def delete_lock(name, resource_group_name=None, resource_provider_namespace=None
     lock_client = _resource_lock_client_factory()
     lock_resource = _validate_lock_params(resource_group_name, resource_provider_namespace,
                                           parent_resource_path, resource_type, resource_name)
+    _validate_lock_params_match_lock(lock_client, name, resource_group_name,
+                                     resource_provider_namespace, parent_resource_path,
+                                     resource_type, resource_name)
+
     resource_group_name = lock_resource[0]
     resource_name = lock_resource[1]
     resource_provider_namespace = lock_resource[2]
     resource_type = lock_resource[3]
+
 
     if resource_group_name is None:
         return lock_client.management_locks.delete_at_subscription_level(name)
@@ -624,7 +715,7 @@ def create_lock(name, resource_group_name=None, resource_provider_namespace=None
     :type notes: str
     '''
     if level != 'ReadOnly' and level != 'CanNotDelete':
-        raise CLIError('--level must be one of "ReadOnly" or "CanNotDelete"')
+        raise CLIError('--lock-type must be one of "ReadOnly" or "CanNotDelete"')
     parameters = ManagementLockObject(level=level, notes=notes, name=name)
 
     lock_client = _resource_lock_client_factory()
@@ -734,10 +825,10 @@ class _ResourceUtils(object): #pylint: disable=too-many-instance-attributes
             else:
                 _validate_resource_inputs(resource_group_name, resource_provider_namespace,
                                           resource_type, resource_name)
-                api_version = _ResourceUtils._resolve_api_version(self.rcf,
-                                                                  resource_provider_namespace,
-                                                                  parent_resource_path,
-                                                                  resource_type)
+                api_version = _ResourceUtils.resolve_api_version(self.rcf,
+                                                                 resource_provider_namespace,
+                                                                 parent_resource_path,
+                                                                 resource_type)
 
         self.resource_group_name = resource_group_name
         self.resource_provider_namespace = resource_provider_namespace
@@ -746,6 +837,34 @@ class _ResourceUtils(object): #pylint: disable=too-many-instance-attributes
         self.resource_name = resource_name
         self.resource_id = resource_id
         self.api_version = api_version
+
+    def create_resource(self, properties, location, is_full_object):
+        res = json.loads(properties)
+        if not is_full_object:
+            if not location:
+                if self.resource_id:
+                    rg_name = parse_resource_id(self.resource_id)['resource_group']
+                else:
+                    rg_name = self.resource_group_name
+                location = self.rcf.resource_groups.get(rg_name).location
+
+            res = GenericResource(location=location, properties=res)
+        elif res.get('location', None) is None:
+            raise IncorrectUsageError("location of the resource is required")
+
+        if self.resource_id:
+            resource = self.rcf.resources.create_or_update_by_id(self.resource_id,
+                                                                 self.api_version,
+                                                                 res)
+        else:
+            resource = self.rcf.resources.create_or_update(self.resource_group_name,
+                                                           self.resource_provider_namespace,
+                                                           self.parent_resource_path or '',
+                                                           self.resource_type,
+                                                           self.resource_name,
+                                                           self.api_version,
+                                                           res)
+        return resource
 
     def get_resource(self):
         if self.resource_id:
@@ -811,7 +930,7 @@ class _ResourceUtils(object): #pylint: disable=too-many-instance-attributes
                 parameters)
 
     @staticmethod
-    def _resolve_api_version(rcf, resource_provider_namespace, parent_resource_path, resource_type):
+    def resolve_api_version(rcf, resource_provider_namespace, parent_resource_path, resource_type):
         provider = rcf.providers.get(resource_provider_namespace)
 
         #If available, we will use parent resource's api-version
@@ -851,4 +970,4 @@ class _ResourceUtils(object): #pylint: disable=too-many-instance-attributes
             parent = None
             resource_type = parts['type']
 
-        return _ResourceUtils._resolve_api_version(rcf, namespace, parent, resource_type)
+        return _ResourceUtils.resolve_api_version(rcf, namespace, parent, resource_type)